Refactor Cloudinary module: Consolidate services and remove redundant interfaces
This commit refactors the Cloudinary module by consolidating the Cloudinary service and removing unnecessary interfaces. The CloudinaryImageService has been eliminated, and its functionality has been integrated into the CloudinaryService. Additionally, several redundant interfaces related to image uploads and transformations have been removed, streamlining the codebase.

- **Changes made:**
  - Merged CloudinaryImageService into CloudinaryService.
  - Removed unused interfaces: CacheEntry, EntityWithImage, UploadedFileType, ImageTransformationOptions, ImageUploadOptions, and EntityRepository.
  - Updated import paths to reflect the new structure.
  - Added tests for the consolidated CloudinaryService.

This refactor improves maintainability and reduces complexity in the Cloudinary module.
