--- src/modules/cloudinary/domain/interfaces/cache-entry.interface.ts
@@ -1,7 +1,0 @@
-/**
- * Interface for cache entries
- */
-export interface CacheEntry {
-  data: unknown;
-  timestamp: number;
-}
