--- src/modules/cloudinary/domain/interfaces/entity-with-image.interface.ts
@@ -1,7 +1,0 @@
-/**
- * Interface for entities that have an image field
- */
-export interface EntityWithImage {
-  id: string;
-  [imageField: string]: any;
-}
