--- src/modules/cloudinary/domain/interfaces/image-transformation-options.interface.ts
@@ -1,7 +1,0 @@
-export interface ImageTransformationOptions {
-  width?: number;
-  height?: number;
-  crop?: 'fill' | 'fit' | 'scale' | 'thumb';
-  quality?: 'auto' | number;
-  format?: 'jpg' | 'png' | 'gif' | 'webp';
-}
