+++ src/modules/cloudinary/types/file-upload.type.ts
@@ -1,0 +1,12 @@
+/**
+ * File upload type for image uploads
+ * Used across the application for consistent file handling
+ */
+export interface UploadedFileType {
+  fieldname: string;
+  originalname: string;
+  encoding: string;
+  mimetype: string;
+  size: number;
+  buffer: Buffer;
+}
