+++ src/modules/cloudinary/types/image-transformation.type.ts
@@ -1,0 +1,11 @@
+/**
+ * Image transformation options for Cloudinary
+ * Used for configuring image transformations during upload
+ */
+export interface ImageTransformationOptions {
+  width?: number;
+  height?: number;
+  crop?: string;
+  quality?: number;
+  format?: string;
+}
